wip: refactor: split out @fadroma/create
- just the project creation tool
- publishing to this NPM package id should enable 'npx fadroma create'
  instead of 'npx @hackbg/fadroma@latest create'
