+++ create/README.md
@@ -1,0 +1,9 @@
+# `@fadroma/create`
+
+This is a command-line tool for creating Fadroma projects.
+
+Invoke with:
+
+```shell
+npx fadroma create
+```
