--- ops/project.docs.md
@@ -1,3 +1,0 @@
-# Fadroma Project Setup
-
-[TODO](https://www.youtube.com/watch?v=VyZiIuMufTA)
